+++ src/styles/Header.module.css
@@ -16,10 +16,9 @@ body{
     flex-direction: row-reverse;
     font-weight: 500;
     box-shadow: 0px 2px 4px 0px rgba(0, 0, 0, 0.25);
-    margin-bottom: 100px;
     left: 0px;
     top: 0px;
-    z-index: 3;
+    z-index: 10;
     background-color: #fff;
 }
 
@@ -71,6 +70,6 @@ body {
 .headerMargin {
     position: relative;
     width: 100%;
-    height: 80px;
+    height: 105px;
 }
