+++ src/styles/product/MergeLayout.module.css
@@ -1,8 +1,9 @@
 .mergeMain {
     position: relative;
     display: flex;
+    justify-content: center;
     align-content: center;
-    width: 1067px;
+    width: 1283px;
     margin: 0 auto;
   }
